pyproject.toml
[tool.poetry]
name = "audioviz"
version = "0.1.0"
description = "Real-time audio visualizer widget for desktop"
authors = ["gephaistos <olehmaksa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23"
numba = "^0.56"
PyGObject = "^3.42"
pycairo = "^1.21"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "^5"
pep8-naming = "*"
autopep8 = "*"
isort = {version = "*", extras = ["colors"]}

[tool.poetry.scripts]
audioviz = "audioviz.cli.run:run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
